pyproject.toml
[project]
name = "agent-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk (>=0.3.0)",
    "googlemaps (>=4.10.0)",
    "python-dotenv (>=1.0.0)",
    "requests (>=2.31.0)",
    "ticketpy (>=1.1.2)",
    "pytz (>=2024.1)",
    "pillow (>=10.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.34.0"

[tool.poe.tasks]
run-agents = "adk web agents"

